# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyTailz API",
    "version": "1.0.0",
    "description": "Local Swagger documentation for HappyTailz backend"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://happytailz-9qf9.onrender.com/", "description": "Render" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pet": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ownerId": { "type": "string" },
          "name": { "type": "string" },
          "species": { "type": "string" },
          "breed": { "type": "string" },
          "age": { "type": "number" },
          "gender": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri", "description": "Direct URL of the pet image" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/pets": {
      "get": {
        "summary": "List my pets",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create pet",
        "tags": ["Pets"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "species"],
                "properties": {
                  "name": { "type": "string" },
                  "species": { "type": "string" },
                  "breed": { "type": "string" },
                  "age": { "type": "number" },
                  "gender": { "type": "string" },
                  "image": { "type": "string", "format": "binary", "description": "Image file upload (field name: image)" },
                  "imageUrl": { "type": "string", "format": "uri", "description": "Optional direct image URL instead of file" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "species"],
                "properties": {
                  "name": { "type": "string" },
                  "species": { "type": "string" },
                  "breed": { "type": "string" },
                  "age": { "type": "number" },
                  "gender": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "summary": "User dashboard",
        "description": "Shows user's latest plan with isActive flag, pets summary, upcoming walking requests, and completed count.",
        "tags": ["Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasPlan": { "type": "boolean" },
                    "plan": { "type": "object", "properties": { "isActive": { "type": "boolean" } } },
                    "pets": { "type": "object", "properties": { "count": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/Pet" } } } },
                    "walkingRequests": { "type": "object", "properties": { "upcoming": { "type": "array" }, "completedCount": { "type": "integer" } } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/userplans": {
      "get": {
        "summary": "List my user plans",
        "tags": ["UserPlans"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Subscribe to a plan",
        "tags": ["UserPlans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["planId"],
                "properties": {
                  "planId": { "type": "string" },
                  "startDate": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/walkingrequests": {
      "post": {
        "summary": "Create walking request",
        "tags": ["WalkingRequests"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["pet", "scheduleAt"],
                "properties": {
                  "pet": { "type": "string" },
                  "plan": { "type": "string" },
                  "scheduleAt": { "type": "string", "format": "date-time" },
                  "location": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/profile": {
      "get": {
        "summary": "Get my profile",
        "tags": ["Profile"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      },
      "put": {
        "summary": "Update my profile",
        "tags": ["Profile"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      }
    },
...
    "/api/admin/breeding": {
      "get": { "summary": "List breeding", "tags": ["Admin - Breeding"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create breeding",
        "tags": ["Admin - Breeding"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/admin/hospitals": {
      "get": { "summary": "List hospitals", "tags": ["Admin - Hospitals"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create hospital",
        "tags": ["Admin - Hospitals"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "address"],
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "address"],
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/admin/grooming": {
      "get": { "summary": "List grooming", "tags": ["Admin - Grooming"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create grooming",
        "tags": ["Admin - Grooming"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": { "type": "string" },
                  "services": { "type": "array", "items": { "type": "string" } },
                  "price": { "type": "number" },
                  "durationMinutes": { "type": "integer" },
                  "active": { "type": "boolean" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/admin/plans": {
      "get": { "summary": "List plans", "tags": ["Admin - Plans"], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create plan",
        "tags": ["Admin - Plans"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "durationDays", "price"],
                "properties": {
                  "name": { "type": "string", "enum": ["1 day", "1 week", "1 month"] },
                  "durationDays": { "type": "integer" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "summary": "List all users",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "driver", "user"] } },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search by name/email" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "name": { "type": "string" },
                      "email": { "type": "string" },
                      "role": { "type": "string" },
                      "activePlan": { "type": "object", "nullable": true }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": ["Admin - Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "image": { "type": "string", "format": "binary" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "driver", "user" ] },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "409": { "description": "Email exists" } }
      }
    },
...
    "/api/admin/dashboard": {
      "get": {
        "summary": "Admin dashboard summary",
        "description": "High-level counts: users by role, pets, active subscriptions, walking requests by status, and recent users.",
        "tags": ["Admin - Dashboard"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}